CCO-xAPP/services/inference/requirements.txt
kafka-python
pandas
scikit-learn
joblib
xgboost
pydantic
